Game-related changes
Moved, deleted, and renamed what are now start, teaser, goldbergAria, and their derivatives.
Removed some other planning and source things to an off-repo area.
Added to README.md.
Remade the link to the game in selector.html.

Did quite a lot to tagame.js:
Changed const to var because why not.
Replaced getMedia with loadResources.
Added many console logs for debugging.
Removed all references to babies, as the game has changed.
Added a cache which stores the images and sounds.