Added some support for system dark mode
This is a really powerful tool that I just discovered, and I will be utilizing this hopefully in the near future. Currently just automatically uses dark theme, but in the future I plan to change navbar and other elements to actually be dark and have good contrast with the text, probably in a new stylesheet that will be imported with a conditional @import.